requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
line-bot-sdk==3.5.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
redis==5.0.1
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
dateparser==1.1.8
spacy==3.7.2
python-dateutil==2.8.2
pytz==2023.3
httpx==0.25.2
pydantic-settings==2.2.1 